Move movie_service class from models folder to services folder and it now recognises env variables
